Revert "Update examples to remove Dagster code from __init__.py (part 2) (#20809)"
This reverts commit 8bd0d1e8617d1d4adefad6543e7e334bdb5faec6.
