+++ examples/assets_dbt_python/assets_dbt_python/assets/dbt/__init__.py
@@ -3,7 +3,7 @@ from typing import Any, Mapping
 from dagster import AssetExecutionContext, AssetKey, file_relative_path
 from dagster_dbt import DagsterDbtTranslator, DbtCliResource, dbt_assets, get_asset_key_for_model
 
-DBT_PROJECT_DIR = file_relative_path(__file__, "../../dbt_project")
+DBT_PROJECT_DIR = file_relative_path(__file__, "../../../dbt_project")
 
 dbt_resource = DbtCliResource(project_dir=DBT_PROJECT_DIR)
 dbt_parse_invocation = dbt_resource.cli(["parse"]).wait()
