+++ examples/assets_dbt_python/pyproject.toml
@@ -3,5 +3,4 @@ requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
 [tool.dagster]
-module_name = "assets_dbt_python.definitions"
-code_location_name = "assets_dbt_python"
+module_name = "assets_dbt_python"
