+++ examples/assets_dynamic_partitions/pyproject.toml
@@ -3,5 +3,4 @@ requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
 [tool.dagster]
-module_name = "assets_dynamic_partitions.definitions"
-code_location_name = "assets_dynamic_partitions"
+module_name = "assets_dynamic_partitions"
