+++ examples/assets_modern_data_stack/assets_modern_data_stack_tests/test_defs.py
@@ -1,4 +1,4 @@
-from assets_modern_data_stack.definitions import defs
+from assets_modern_data_stack import defs
 
 
 def test_defs_can_load():
