+++ examples/assets_modern_data_stack/pyproject.toml
@@ -3,5 +3,4 @@ requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
 [tool.dagster]
-module_name = "assets_modern_data_stack.definitions"
-code_location_name = "assets_modern_data_stack"
+module_name = "assets_modern_data_stack"
