+++ examples/assets_pandas_pyspark/assets_pandas_pyspark_tests/test_defs.py
@@ -1,4 +1,4 @@
-from assets_pandas_pyspark.definitions import get_spark_weather_defs, get_weather_defs
+from assets_pandas_pyspark import get_spark_weather_defs, get_weather_defs
 
 
 def test_weather_defs_can_load():
