+++ examples/assets_pandas_pyspark/pyproject.toml
@@ -3,5 +3,4 @@ requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
 [tool.dagster]
-module_name = "assets_pandas_pyspark.definitions"
-code_location_name = "assets_pandas_pyspark"
+module_name = "assets_pandas_pyspark"
