+++ examples/assets_smoke_test/assets_smoke_test/pure_python_assets.py
