+++ examples/assets_smoke_test/assets_smoke_test_tests/test_smoke_pure_python_assets.py
@@ -1,7 +1,7 @@
 from dagster import InMemoryIOManager, TableSchema, load_assets_from_modules, materialize
 from pandas import DataFrame, Series
 
-from assets_smoke_test.assets import pure_python_assets
+from assets_smoke_test import pure_python_assets
 
 
 def empty_dataframe_from_column_schema(column_schema: TableSchema) -> DataFrame:
