+++ examples/assets_smoke_test/assets_smoke_test_tests/test_smoke_python_and_dbt_assets.py
@@ -5,8 +5,8 @@ from dagster import load_assets_from_modules, materialize
 from dagster_dbt import DbtCliResource
 from dagster_snowflake_pandas import SnowflakePandasIOManager
 
-from assets_smoke_test.assets import python_and_dbt_assets
-from assets_smoke_test.assets.python_and_dbt_assets import DBT_PROJECT_DIR, raw_country_populations
+from assets_smoke_test import python_and_dbt_assets
+from assets_smoke_test.python_and_dbt_assets import DBT_PROJECT_DIR, raw_country_populations
 
 
 def smoke_all_test():
