+++ examples/experimental/assets_yaml_dsl/pyproject.toml
@@ -3,5 +3,4 @@ requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
 [tool.dagster]
-module_name = "assets_yaml_dsl.definitions"
-code_location_name = "assets_yaml_dsl"
+module_name = "assets_yaml_dsl"
