+++ examples/experimental/external_assets/workspace.yaml
@@ -1,4 +1,2 @@
 load_from:
-  - python_package:
-      package_name: external_assets.definitions
-      location_name: external_assets
+  - python_package: external_assets
