+++ examples/experimental/sling_decorator/pyproject.toml
@@ -12,5 +12,4 @@ dependencies = [
 packages = ["sling_decorator"]
 
 [tool.dagster]
-module_name = "sling_decorator.definitions"
-code_location_name = "sling_decorator"
+module_name = "sling_decorator"
