+++ examples/feature_graph_backed_assets/feature_graph_backed_assets/definitions.py
@@ -9,6 +9,11 @@ from dagster import (
 from . import assets
 from .graphs_and_ops import layover_breakdown_2022, us_assets
 
+airline_job = define_asset_job(
+    "airline_job", AssetSelection.assets("passenger_flights").downstream()
+)
+
+
 defs = Definitions(
     assets=[
         *load_assets_from_package_module(assets),
